+++ cli/tests/test_rebase_command.rs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 use crate::common::create_commit;
+use crate::common::create_commit_with_files;
 use crate::common::CommandOutput;
 use crate::common::TestEnvironment;
 use crate::common::TestWorkDir;
@@ -2963,6 +2964,70 @@ fn test_rebase_skip_if_on_destination() {
     ○ │  b1  zsuskuln  62634b59:  a
     ├─╯
     ○  a  rlvkpnrz  7d980be7
+    ◆    zzzzzzzz  00000000
+    [EOF]
+    ");
+}
+
+#[test]
+fn test_rebase_skip_duplicate_divergent() {
+    let test_env = TestEnvironment::default();
+    test_env.run_jj_in(".", ["git", "init", "repo"]).success();
+    let work_dir = test_env.work_dir("repo");
+
+    // Set up commit graph with divergent changes
+    create_commit_with_files(&work_dir, "a", &[], &[("file1", "initial\n")]);
+    create_commit_with_files(&work_dir, "b2", &["a"], &[("file1", "initial\nb\n")]);
+    create_commit_with_files(&work_dir, "c", &["a"], &[("file2", "d\n")]);
+    work_dir.run_jj(["rebase", "-s", "b2", "-d", "c"]).success();
+    work_dir
+        .run_jj(["bookmark", "create", "b1", "-r", "at_operation(@-, b2)"])
+        .success();
+
+    // Test the setup (commit B is duplicated)
+    insta::assert_snapshot!(get_long_log_output(&work_dir), @r"
+    ○  b2  zsuskuln  3bb70fa2:  c
+    @  c  royxmykx  ddc47df6:  a
+    │ ○  b1  zsuskuln  48bf33ab:  a
+    ├─╯
+    ○  a  rlvkpnrz  08789390
+    ◆    zzzzzzzz  00000000
+    [EOF]
+    ");
+
+    // By default, rebase should skip the duplicate of commit B
+    insta::assert_snapshot!(work_dir.run_jj(["rebase", "-s", "c", "-d", "b1"]).success(), @r"
+    ------- stderr -------
+    Skipped 1 divergent commits that were already present in the destination
+    Rebased 1 commits to destination
+    Working copy  (@) now at: royxmykx e09e76dc b2 c | c
+    Parent commit (@-)      : zsuskuln 48bf33ab b1 | b2
+    Added 0 files, modified 1 files, removed 0 files
+    [EOF]
+    ");
+    insta::assert_snapshot!(get_long_log_output(&work_dir), @r"
+    @  b2 c  royxmykx  e09e76dc:  b1
+    ○  b1  zsuskuln  48bf33ab:  a
+    ○  a  rlvkpnrz  08789390
+    ◆    zzzzzzzz  00000000
+    [EOF]
+    ");
+
+    // Rebase with "--keep-divergent" shouldn't skip any duplicates
+    work_dir.run_jj(["undo"]).success();
+    insta::assert_snapshot!(work_dir.run_jj(["rebase", "-s", "c", "-d", "b1", "--keep-divergent"]).success(), @r"
+    ------- stderr -------
+    Rebased 2 commits to destination
+    Working copy  (@) now at: royxmykx 1e2f1d7b c | c
+    Parent commit (@-)      : zsuskuln?? 48bf33ab b1 | b2
+    Added 0 files, modified 1 files, removed 0 files
+    [EOF]
+    ");
+    insta::assert_snapshot!(get_long_log_output(&work_dir), @r"
+    ○  b2  zsuskuln  0f72e47f:  c
+    @  c  royxmykx  1e2f1d7b:  b1
+    ○  b1  zsuskuln  48bf33ab:  a
+    ○  a  rlvkpnrz  08789390
     ◆    zzzzzzzz  00000000
     [EOF]
     ");